fix #13: 분석 리스너 두개를 하나로 합침
